Code refactoring
