+++ src/countyData/candidatesInfo/utils.ts
@@ -1,0 +1,11 @@
+export const isBidenWins = (biden: number, trump: number) => {
+  return biden > trump ? true : false;
+};
+
+export const getColor = (biden: boolean | undefined, opacity: number) => {
+  if (biden) {
+    return `rgba(52, 125, 192, ${opacity})`;
+  } else {
+    return `rgba(231, 33, 33, ${opacity})`;
+  }
+};
